[update] Injetando usuário nas requisições para relacionamento com tarefas

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,40 +14,15 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    // getAllTasks(): Task[] {
-    //     return this.tasks;
-    // }
-
-    // getTasksWithFilters(filterDTO: GetTasksFilterDTO): Task[]{
-    //     const { status, search } = filterDTO;
-    //     let tasks = this.getAllTasks();
-
-    //     if (status) {
-    //         tasks = tasks.filter(task => task.status === status);
-    //     }
-    //     if (search){
-    //         tasks = tasks.filter(task => 
-    //             task.title.includes(search) || task.description.includes(search));
-    //     }
-
-    //     return tasks;
-    // }
-
-    async createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
-        const { title, description } = createTaskDTO;
-
-        const task = new Task();
-        task.title = title;
-        task.description = description;
-        task.status = TaskStatus.OPEN;
-
-        await task.save();
-
-        return task;
+    async createTask(
+        createTaskDTO: CreateTaskDTO, 
+        user: User
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDTO, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task com id ${id} não encontrado.`);
@@ -55,17 +31,20 @@ export class TasksService {
         return found;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task com id ${id} não encontrado.`);
         }
-        // return result;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User
+    ): Promise<Task>{
+        const task = await this.getTaskById(id, user);
 
         task.status = status;
         await task.save();
@@ -73,7 +52,7 @@ export class TasksService {
         return task;
     }
 
-    async getTasks(filterDto: GetTasksFilterDTO): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTasksFilterDTO, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 }
